+++ functions/PreparePHP_SELF.fnc.php
@@ -274,3 +274,56 @@ function _myURLEncode( $string )
 {
 	return URLEscape( $string );
 }
+
+/**
+ * RosarioSIS URL: dir (site), or script (page), or request (includes request params)
+ * Will detect https inside Docker or behind reverse proxy
+ *
+ * Given this URL: https://domain.com/rosariosis/Modules.php?modname=Module/Program.php
+ * @example dir: https://domain.com/rosariosis/ (with trailing slash)
+ * @example script: https://domain.com/rosariosis/Modules.php
+ * @example request: https://domain.com/rosariosis/Modules.php?modname=Module/Program.php
+ *
+ * @since 11.2
+ *
+ * @param string $mode dir (site), or script (page), or request (include request params).
+ *
+ * @return string Escaped URL.
+ */
+function RosarioURL( $mode = 'dir' )
+{
+	$url = 'http://';
+
+	if ( ( isset( $_SERVER['HTTPS'] ) && $_SERVER['HTTPS'] === 'on' )
+		|| ( isset( $_SERVER['HTTP_X_FORWARDED_PROTO'] ) && $_SERVER['HTTP_X_FORWARDED_PROTO'] === 'https' )
+		|| ( isset( $_SERVER['HTTP_X_FORWARDED_SSL'] ) && $_SERVER['HTTP_X_FORWARDED_SSL'] === 'on' ) )
+	{
+		// Fix detect https inside Docker or behind reverse proxy.
+		$url = 'https://';
+	}
+
+	$url .= $_SERVER['SERVER_NAME'];
+
+	if ( $_SERVER['SERVER_PORT'] != '80'
+		&& $_SERVER['SERVER_PORT'] != '443' )
+	{
+		$url .= ':' . $_SERVER['SERVER_PORT'];
+	}
+
+	if ( $mode === 'dir' )
+	{
+		$url .= dirname( $_SERVER['SCRIPT_NAME'] ) === DIRECTORY_SEPARATOR ?
+			// Add trailing slash.
+			'/' : dirname( $_SERVER['SCRIPT_NAME'] ) . '/';
+	}
+	elseif ( $mode === 'script' )
+	{
+		$url .= $_SERVER['SCRIPT_NAME'];
+	}
+	elseif ( $mode === 'request' )
+	{
+		$url .= $_SERVER['REQUEST_URI'];
+	}
+
+	return URLEscape( $url );
+}